refactor: combine usetiful entry file, create tour container component
