+++ src/Survey.tsx
@@ -6,9 +6,11 @@ const Survey = ({ survey }: { survey: SurveyType }) => {
   //TODO implement survey
   console.log('survey', survey);
   return (
-    <View style={styles.modal}>
-      <StepHeader {...{ title: 'placeholder' }} />
-      <Text>Survey</Text>
+    <View style={styles.container}>
+      <View style={styles.modal}>
+        <StepHeader {...{ title: 'placeholder' }} />
+        <Text>Survey</Text>
+      </View>
     </View>
   );
 };
@@ -16,6 +18,12 @@ const Survey = ({ survey }: { survey: SurveyType }) => {
 export default Survey;
 
 const styles = StyleSheet.create({
+  container: {
+    position: 'absolute',
+    width: '100%',
+    height: '100%',
+    backgroundColor: '#000000cc',
+  },
   modal: {
     backgroundColor: 'white',
     marginTop: '50%',
