+++ src/index.ts
@@ -1,3 +1,2 @@
 export { Usetiful } from './Usetiful';
 export { setPointer } from './utils/setPointer';
-export { UsetifulTemp } from './UsetifulTemp';
